Made an outline shadergraph, a new UI prototype, and added mouse lock to player script (sry joe)
I followed a quick tutorial and made a simple shadergraph material that creates an outline on any object on the Outline layer. It's mostly just for testing purposes for now. I also made a super simple script for an enemy placeholder that highlights it when looked at. I created a new UI prototype for script stealing that works with the new system of having 1 script slot for movement and 1 for combat. I also added mouse lock to the player movement script sry joe lol