Added device detection for UI and continued script steal prototype
I added functionality in the UI script stealing stuff to detect if a controller or keyboard is being used, and it changes whether the mouse is unlocked and followed by the little dot thing or if it's controlled by the controller's joystick. I also started working on the Hud reflect stealing a script to get a feel for what feels best control-wise.