Boss Version 1
Started working on the boss fight, for now I have a simple keyboard model and a basic animation. I was thinking the keyboard could type out what was going to happen.... even if the player might not ever see it. For now if the "isTesting" bool is checked and the p key is pressed the keyboard will type out Fall and a projectile will spawn in the air and fall down to hit the player. Will play around with this some more tommorow.