Script stealing applies to all enemies with that script, basic enemy health added
I updated the script stealing to apply to all enemies that have that script type when a script is stolen, and added basic enemy health to the enemies. Currently there isn't a way to attack the enemies since no attack function on the player is linked to any buttons I know of.