Spite-Bulb Reflects if Script is Stolen
When the electricity script is stolen, the spite bulb color goes from bright yellow to a dull gray, and vise-versa when the script is returned. Also added ability to "return" scripts manually without replacing them, done by pressing X on controller or Q on keyboard when selecting a script slot.