Update to Support Health Classes and Targeting Priority
Targetable Objects (things like enemies and destructible objects like pots) can now be assigned a TargetingWeight which determines it's priority in the auto-targeting process during combat. For example, we wouldn't ever want the player to target a breakable pot over an enemy, so we reduce the TargetWeight for the pot to 0.2f.

Updated Player Combat controller to make auto targeting more accurate in terms of how close the player needs to push the joystick towards an enemy to be able to dash to it.